+++ models/Place.go
@@ -1,7 +1,7 @@
 package main
 
 type Place struct {
-	Id   uint16  `json:"id"`
+	Id   int     `json:"id"`
 	Name string  `json:"name"`
 	Lat  float64 `json:"lat"`
 	Lng  float64 `json:"lng"`
