+++ models/eppi/Volume.go
@@ -1,8 +1,8 @@
 package main
 
 type Volume struct {
-	Id   uint16 `json:"id"`
-	Name byte   `json:"name"`
+	Id   int  `json:"id"`
+	Name byte `json:"name"`
 }
 
 type Volumes []Volume
